+++ lib/rails_state_machine/model.rb
@@ -36,8 +36,11 @@ module RailsStateMachine
     private
 
     def state_machine_state_manager(state_attribute)
-      @state_machine_state_managers ||= {}
       @state_machine_state_managers[state_attribute] ||= StateManager.new(self, state_machine(state_attribute), state_attribute)
+    end
+
+    def clear_state_machine_state_managers_cache
+      @state_machine_state_managers = {}
     end
 
     def state_machine_state_managers