+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Unreleased
 
 ### Compatible changes
+- Fix bug where duplicated model instances referenced shared state machine state managers.
 
 ### Breaking changes